finished delete middleware by category

+++ middleware/delete.middleware.js
@@ -35,8 +35,8 @@ module.exports = {
         id: req.params.id
       },
       limit: 1
-    }).then((responseData) => {
-      res.json(responseData)
+    }).then(() => {
+      res.json({ 'success': 'success!' })
     }).
     catch((err) => {
       console.error(err)
@@ -45,11 +45,39 @@ module.exports = {
   },
   deleteCategory: (req, res) => {
     console.log(req.params)
-    Project.destroy({
+    const responseObject = {}
+    Project.findAll({
       where: {
         category: req.params.category
       }
     }).then((responseData) => {
+      for (const ind in responseData) {
+        if (Object.hasOwnProperty.call(responseData, ind)) {
+          const projectId = responseData[ind].dataValues.id
+          Blob.destroy({
+            where: {
+              projectId
+            }
+          }).then((responseBlob) => {
+            Project.destroy({
+              where: {
+                id: projectId
+              }
+            }).then((responseProject) => {
+              responseObject[`Blobs with projectId ${projectId}`] = responseBlob
+              responseObject[`Projects with id ${projectId}`] = responseProject
+            }).
+            catch((err) => {
+              console.error(err)
+              res.json(err)
+            })
+          }).
+          catch((err) => {
+            console.error(err)
+            res.json(err)
+          })
+        }
+      }
       res.json(responseData)
     }).
     catch((err) => {
